+++ CHANGELOG.md
@@ -1,1 +1,4 @@
-Initial Release
+### 0.2.0
+
+- Added regex support to advanced search
+- Added missing lang entries